refactor(comics): extract PanelMaskBinarizer from PanelDetector
binarizeMask() belongs to report generation, not the detection algorithm.
Move it to a dedicated PanelMaskBinarizer object with its own test file.
Extract detectBackgroundLuma() as a package-internal detectPageBackground()
function shared by both PanelDetector and PanelMaskBinarizer.
